+++ cardano-cli/src/Cardano/CLI/Legacy/Commands/Governance.hs
@@ -35,19 +35,6 @@ data LegacyGovernanceCmds
       [VerificationKeyFile In]
       ProtocolParametersUpdate
       (Maybe FilePath)
-  | GovernanceCreatePoll
-      Text -- Prompt
-      [Text] -- Choices
-      (Maybe Word) -- Nonce
-      (File GovernancePoll Out)
-  | GovernanceAnswerPoll
-      (File GovernancePoll In) -- Poll file
-      (Maybe Word) -- Answer index
-      (Maybe (File () Out)) -- Tx file
-  | GovernanceVerifyPoll
-      (File GovernancePoll In) -- Poll file
-      (File (Tx ()) In) -- Tx file
-      (Maybe (File () Out)) -- Tx file
   deriving Show
 
 renderLegacyGovernanceCmds :: LegacyGovernanceCmds -> Text
@@ -57,7 +44,4 @@ renderLegacyGovernanceCmds = \case
   GovernanceMIRTransfer _ _ _ TransferToTreasury -> "governance create-mir-certificate transfer-to-treasury"
   GovernanceMIRTransfer _ _ _ TransferToReserves -> "governance create-mir-certificate transfer-to-reserves"
   GovernanceUpdateProposal {} -> "governance create-update-proposal"
-  GovernanceCreatePoll{} -> "governance create-poll"
-  GovernanceAnswerPoll{} -> "governance answer-poll"
-  GovernanceVerifyPoll{} -> "governance verify-poll"
