+++ cardano-cli/src/Cardano/CLI/Legacy/Run/Governance.hs
@@ -1,7 +1,6 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
 module Cardano.CLI.Legacy.Run.Governance
@@ -11,32 +10,20 @@ module Cardano.CLI.Legacy.Run.Governance
 import           Cardano.Api
 import qualified Cardano.Api.Ledger as Ledger
 import           Cardano.Api.Shelley
-import qualified Cardano.Api.Shelley as Api
 
 import           Cardano.CLI.EraBased.Run.Governance
 import           Cardano.CLI.Legacy.Commands.Governance
-import           Cardano.CLI.Read
 import           Cardano.CLI.Types.Common
 import           Cardano.CLI.Types.Errors.GovernanceCmdError
 import           Cardano.CLI.Types.Key
 
 import           Control.Monad
-import           Control.Monad.IO.Class
-import           Control.Monad.Trans.Class (lift)
 import           Control.Monad.Trans.Except (ExceptT)
 import           Control.Monad.Trans.Except.Extra
 import           Data.Aeson (eitherDecode)
-import qualified Data.ByteString.Char8 as BSC
 import qualified Data.ByteString.Lazy as LB
 import           Data.Function ((&))
-import           Data.String (fromString)
-import           Data.Text (Text)
 import qualified Data.Text as Text
-import qualified Data.Text.Encoding as Text
-import qualified Data.Text.IO as Text
-import qualified Data.Text.Read as Text
-import qualified System.IO as IO
-import           System.IO (stderr, stdin, stdout)
 
 runLegacyGovernanceCmds :: LegacyGovernanceCmds -> ExceptT GovernanceCmdError IO ()
 runLegacyGovernanceCmds = \case
@@ -48,12 +35,6 @@ runLegacyGovernanceCmds = \case
     runLegacyGovernanceGenesisKeyDelegationCertificate sbe genVk genDelegVk vrfVk out
   GovernanceUpdateProposal out eNo genVKeys ppUp mCostModelFp ->
     runLegacyGovernanceUpdateProposal out eNo genVKeys ppUp mCostModelFp
-  GovernanceCreatePoll prompt choices nonce out ->
-    runLegacyGovernanceCreatePoll prompt choices nonce out
-  GovernanceAnswerPoll poll ix mOutFile ->
-    runLegacyGovernanceAnswerPoll poll ix mOutFile
-  GovernanceVerifyPoll poll metadata mOutFile ->
-    runLegacyGovernanceVerifyPoll poll metadata mOutFile
 
 runLegacyGovernanceMIRCertificatePayStakeAddrs
   :: EraInEon ShelleyToBabbageEra
@@ -163,128 +144,3 @@ runLegacyGovernanceUpdateProposal upFile eNo genVerKeyFiles upPprams mCostModelF
 
   firstExceptT GovernanceCmdTextEnvWriteError . newExceptT
     $ writeLazyByteStringFile upFile $ textEnvelopeToJSON Nothing upProp
-
-runLegacyGovernanceCreatePoll
-  :: Text
-  -> [Text]
-  -> Maybe Word
-  -> File GovernancePoll Out
-  -> ExceptT GovernanceCmdError IO ()
-runLegacyGovernanceCreatePoll govPollQuestion govPollAnswers govPollNonce out = do
-  let poll = GovernancePoll{ govPollQuestion, govPollAnswers, govPollNonce }
-
-  let description = fromString $ "An on-chain poll for SPOs: " <> Text.unpack govPollQuestion
-  firstExceptT GovernanceCmdTextEnvWriteError . newExceptT $
-    writeFileTextEnvelope out (Just description) poll
-
-  let metadata = asTxMetadata poll
-        & metadataToJson TxMetadataJsonDetailedSchema
-
-  let outPath = unFile out & Text.encodeUtf8 . Text.pack
-
-  liftIO $ do
-    BSC.hPutStrLn stderr $ mconcat
-      [ "Poll created successfully.\n"
-      , "Please submit a transaction using the resulting metadata.\n"
-      ]
-    BSC.hPutStrLn stdout (prettyPrintJSON metadata)
-    BSC.hPutStrLn stderr $ mconcat
-      [ "\n"
-      , "Hint (1): Use '--json-metadata-detailed-schema' and '--metadata-json-file' "
-      , "from the build or build-raw commands.\n"
-      , "Hint (2): You can redirect the standard output of this command to a JSON "
-      , "file to capture metadata.\n\n"
-      , "Note: A serialized version of the poll suitable for sharing with "
-      , "participants has been generated at '" <> outPath <> "'."
-      ]
-
-runLegacyGovernanceAnswerPoll
-  :: File GovernancePoll In
-  -> Maybe Word -- ^ Answer index
-  -> Maybe (File () Out) -- ^ Output file
-  -> ExceptT GovernanceCmdError IO ()
-runLegacyGovernanceAnswerPoll pollFile maybeChoice mOutFile = do
-  poll <- firstExceptT GovernanceCmdTextEnvReadError . newExceptT $
-    readFileTextEnvelope AsGovernancePoll pollFile
-
-  choice <- case maybeChoice of
-    Nothing -> do
-      askInteractively poll
-    Just ix -> do
-      validateChoice poll ix
-      liftIO $ BSC.hPutStrLn stderr $ Text.encodeUtf8 $ Text.intercalate "\n"
-        [ govPollQuestion poll
-        , "→ " <> (govPollAnswers poll !! fromIntegral ix)
-        , ""
-        ]
-      pure ix
-
-  let pollAnswer = GovernancePollAnswer
-        { govAnsPoll = hashGovernancePoll poll
-        , govAnsChoice = choice
-        }
-  let metadata =
-        metadataToJson TxMetadataJsonDetailedSchema (asTxMetadata pollAnswer)
-
-  liftIO $ BSC.hPutStrLn stderr $ mconcat
-      [ "Poll answer created successfully.\n"
-      , "Please submit a transaction using the resulting metadata.\n"
-      , "To be valid, the transaction must also be signed using a valid key\n"
-      , "identifying your stake pool (e.g. your cold key).\n"
-      ]
-
-  lift (writeByteStringOutput mOutFile (prettyPrintJSON metadata))
-    & onLeft (left . GovernanceCmdWriteFileError)
-
-  liftIO $ BSC.hPutStrLn stderr $ mconcat
-      [ "\n"
-      , "Hint (1): Use '--json-metadata-detailed-schema' and '--metadata-json-file' "
-      , "from the build or build-raw commands.\n"
-      , "Hint (2): You can redirect the standard output of this command to a JSON "
-      , "file to capture metadata."
-      ]
- where
-  validateChoice :: GovernancePoll -> Word -> ExceptT GovernanceCmdError IO ()
-  validateChoice GovernancePoll{govPollAnswers} ix = do
-    let maxAnswerIndex = length govPollAnswers - 1
-    when (fromIntegral ix > maxAnswerIndex) $ left $
-      GovernanceCmdPollOutOfBoundAnswer maxAnswerIndex
-
-  askInteractively :: GovernancePoll -> ExceptT GovernanceCmdError IO Word
-  askInteractively poll@GovernancePoll{govPollQuestion, govPollAnswers} = do
-    liftIO $ BSC.hPutStrLn stderr $ Text.encodeUtf8 $ Text.intercalate "\n"
-      ( govPollQuestion
-      : [ "[" <> textShow ix <> "] " <> answer
-        | (ix :: Int, answer) <- zip [0..] govPollAnswers
-        ]
-      )
-    liftIO $ BSC.hPutStrLn stderr ""
-    liftIO $ BSC.hPutStr stderr "Please indicate an answer (by index): "
-    txt <- liftIO $ Text.hGetLine stdin
-    liftIO $ BSC.hPutStrLn stderr ""
-    case Text.decimal txt of
-      Right (choice, rest) | Text.null rest ->
-        choice <$ validateChoice poll choice
-      _ ->
-        left GovernanceCmdPollInvalidChoice
-
-runLegacyGovernanceVerifyPoll
-  :: File GovernancePoll In
-  -> File (Api.Tx ()) In
-  -> Maybe (File () Out) -- ^ Output file
-  -> ExceptT GovernanceCmdError IO ()
-runLegacyGovernanceVerifyPoll pollFile txFile mOutFile = do
-  poll <- firstExceptT GovernanceCmdTextEnvReadError . newExceptT $
-    readFileTextEnvelope AsGovernancePoll pollFile
-
-  txFileOrPipe <- liftIO $ fileOrPipe (unFile txFile)
-  tx <- firstExceptT GovernanceCmdCddlError . newExceptT $
-    readFileTx txFileOrPipe
-
-  signatories <- firstExceptT GovernanceCmdVerifyPollError . newExceptT $ pure $
-    verifyPollAnswer poll tx
-
-  liftIO $ IO.hPutStrLn stderr $ "Found valid poll answer with " <> show (length signatories) <> " signatories"
-
-  lift (writeByteStringOutput mOutFile (prettyPrintJSON signatories))
-    & onLeft (left . GovernanceCmdWriteFileError)
